+++ commons-jms-utils/src/main/java/co/com/bancolombia/commons/jms/internal/listener/MQContextListener.java
@@ -9,8 +9,10 @@ import jakarta.jms.ConnectionFactory;
 import jakarta.jms.Destination;
 import jakarta.jms.JMSConsumer;
 import jakarta.jms.JMSContext;
+import jakarta.jms.JMSException;
 import jakarta.jms.MessageListener;
 import jakarta.jms.Queue;
+import jakarta.jms.TemporaryQueue;
 import lombok.experimental.SuperBuilder;
 import lombok.extern.log4j.Log4j2;
 
@@ -22,19 +24,29 @@ public class MQContextListener extends AbstractJMSReconnectable<MQContextListene
     private final MQListenerConfig config;
     private final MQQueuesContainer container;
     private final MQBrokerUtils utils;
+    private final boolean temporary;
     private JMSConsumer consumer;
     private JMSContext context;
+    private TemporaryQueue tempQueue;
 
     @Override
     protected String name() {
         String[] parts = Thread.currentThread().getName().split("-");
-        String finalName = "mq-listener-fixed-queue-" + parts[parts.length - 1] + "[" + config.getQueue() + "]";
+        String finalName;
+        if (temporary) {
+            finalName = "mq-listener-tmp-queue-" + parts[parts.length - 1] + "[" + config.getTempQueueAlias() + "]";
+        } else {
+            finalName = "mq-listener-fixed-queue-" + parts[parts.length - 1] + "[" + config.getQueue() + "]";
+        }
         Thread.currentThread().setName(finalName);
         return finalName;
     }
 
     @Override
     protected void disconnect() {
+        if (temporary) {
+            container.unregisterFromQueueGroup(config.getTempQueueAlias(), tempQueue);
+        }
         if (consumer != null) {
             try {
                 consumer.close();
@@ -55,13 +67,30 @@ public class MQContextListener extends AbstractJMSReconnectable<MQContextListene
     protected MQContextListener connect() {
         log.info("Starting listener {}", getProcess());
         context = connectionFactory.createContext();
-        Destination destination = MQQueueUtils.setupFixedQueue(context, config);
-        consumer = context.createConsumer(destination);//NOSONAR
-        container.registerQueue(config.getQueue(), (Queue) destination);
-        utils.setQueueManager(context, (Queue) destination);
-        consumer.setMessageListener(listener);
         context.setExceptionListener(this);
-        log.info("Listener {} started successfully", getProcess());
+        if (temporary) {
+            tempQueue = MQQueueUtils.setupTemporaryQueue(context, config);
+            container.registerToQueueGroup(config.getTempQueueAlias(), tempQueue);
+            consumer = context.createConsumer(tempQueue);//NOSONAR
+            consumer.setMessageListener(listener);
+            log.info("Listener {} started successfully with queue: {}", getProcess(), shortDestinationName());
+        } else {
+            Destination destination = MQQueueUtils.setupFixedQueue(context, config);
+            utils.setQueueManager(context, (Queue) destination);
+            container.registerQueue(config.getQueue(), (Queue) destination);
+            consumer = context.createConsumer(destination);//NOSONAR
+            consumer.setMessageListener(listener);
+            log.info("Listener {} started successfully with queue: {}", getProcess(), config.getQueue());
+        }
         return this;
+    }
+
+    private String shortDestinationName() {
+        try {
+            return tempQueue.getQueueName().split("\\?")[0];
+        } catch (JMSException e) {
+            log.warn("Error getting temp queue name", e);
+            return "Error getting queue name";
+        }
     }
 }
