+++ commons-jms-utils/src/main/java/co/com/bancolombia/commons/jms/utils/MQMessageListenerUtils.java
@@ -4,7 +4,6 @@ import co.com.bancolombia.commons.jms.api.MQBrokerUtils;
 import co.com.bancolombia.commons.jms.api.MQQueuesContainer;
 import co.com.bancolombia.commons.jms.api.exceptions.MQHealthListener;
 import co.com.bancolombia.commons.jms.internal.listener.MQContextListener;
-import co.com.bancolombia.commons.jms.internal.listener.MQContextTemporaryListener;
 import co.com.bancolombia.commons.jms.internal.models.MQListenerConfig;
 import co.com.bancolombia.commons.jms.internal.models.RetryableConfig;
 import jakarta.jms.ConnectionFactory;
@@ -32,24 +31,11 @@ public class MQMessageListenerUtils {
         if (log.isInfoEnabled()) {
             log.info("Creating {} listeners", config.getConcurrency());
         }
-        if (StringUtils.isNotBlank(config.getTempQueueAlias())) {
-            createListenersTemp(cf, listener, container, config, healthListener, retryableConfig);
-        } else {
-            createListenersFixed(cf, listener, container, utils, config, healthListener, retryableConfig);
-        }
-    }
-
-    private static void createListenersFixed(ConnectionFactory cf,
-                                             MessageListener listener,
-                                             MQQueuesContainer container,
-                                             MQBrokerUtils utils,
-                                             MQListenerConfig config,
-                                             MQHealthListener healthListener,
-                                             RetryableConfig retryableConfig) {
         ExecutorService service = Executors.newFixedThreadPool(config.getConcurrency());
         IntStream.range(0, config.getConcurrency())
                 .mapToObj(number -> MQContextListener.builder()
                         .connectionFactory(cf)
+                        .temporary(StringUtils.isNotBlank(config.getTempQueueAlias()))
                         .utils(utils)
                         .config(config)
                         .listener(listener)
@@ -60,28 +46,6 @@ public class MQMessageListenerUtils {
                 .forEach(service::submit);
         if (log.isInfoEnabled()) {
             log.info("{} listeners created for {}", config.getConcurrency(), config.getQueue());
-        }
-    }
-
-    private static void createListenersTemp(ConnectionFactory cf,
-                                            MessageListener listener,
-                                            MQQueuesContainer container,
-                                            MQListenerConfig config,
-                                            MQHealthListener healthListener,
-                                            RetryableConfig retryableConfig) {
-        ExecutorService service = Executors.newFixedThreadPool(config.getConcurrency());
-        IntStream.range(0, config.getConcurrency())
-                .mapToObj(number -> MQContextTemporaryListener.builder()
-                        .connectionFactory(cf)
-                        .config(config)
-                        .listener(listener)
-                        .container(container)
-                        .healthListener(healthListener)
-                        .retryableConfig(retryableConfig)
-                        .build())
-                .forEach(service::submit);
-        if (log.isInfoEnabled()) {
-            log.info("{} listeners created for {}", config.getConcurrency(), config.getTempQueueAlias());
         }
     }
